feat: refactor getOauthProviderUrl() for reusability
