+++ account-kit/signer/src/client/index.ts
@@ -3,9 +3,8 @@ import { getWebAuthnAttestation } from "@turnkey/http";
 import { IframeStamper } from "@turnkey/iframe-stamper";
 import { WebauthnStamper } from "@turnkey/webauthn-stamper";
 import { z } from "zod";
-import { OAuthProvidersError } from "../errors.js";
-import { getDefaultScopeAndClaims, getOauthNonce } from "../oauth.js";
-import type { AuthParams, OauthMode } from "../signer.js";
+import { getOauthNonce } from "../oauth.js";
+import type { AuthParams } from "../signer.js";
 import { base64UrlEncode } from "../utils/base64UrlEncode.js";
 import { generateRandomBuffer } from "../utils/generateRandomBuffer.js";
 import { BaseSignerClient } from "./base.js";
@@ -17,7 +16,6 @@ import type {
   EmailAuthParams,
   ExportWalletParams,
   OauthConfig,
-  OauthParams,
   OtpParams,
   User,
 } from "./types.js";
@@ -45,16 +43,6 @@ export const AlchemySignerClientParamsSchema = z.object({
 export type AlchemySignerClientParams = z.input<
   typeof AlchemySignerClientParamsSchema
 >;
-
-type OauthState = {
-  authProviderId: string;
-  isCustomProvider?: boolean;
-  requestKey: string;
-  turnkeyPublicKey: string;
-  expirationSeconds?: number;
-  redirectUrl?: string;
-  openerOrigin?: string;
-};
 
 /**
  * A lower level client used by the AlchemySigner used to communicate with
@@ -462,7 +450,15 @@ export class AlchemySignerWebClient extends BaseSignerClient<ExportWalletParams>
   public override oauthWithRedirect = async (
     args: Extract<AuthParams, { type: "oauth"; mode: "redirect" }>
   ): Promise<never> => {
-    const providerUrl = await this.getOauthProviderUrl(args);
+    const turnkeyPublicKey = await this.initIframeStamper();
+
+    const oauthParams = args;
+    const providerUrl = await this.getOauthProviderUrl({
+      oauthParams,
+      turnkeyPublicKey,
+      oauthCallbackUrl: this.oauthCallbackUrl,
+    });
+
     window.location.href = providerUrl;
     return new Promise((_, reject) =>
       setTimeout(() => reject("Failed to redirect to OAuth provider"), 1000)
@@ -498,7 +494,13 @@ export class AlchemySignerWebClient extends BaseSignerClient<ExportWalletParams>
   public override oauthWithPopup = async (
     args: Extract<AuthParams, { type: "oauth"; mode: "popup" }>
   ): Promise<User> => {
-    const providerUrl = await this.getOauthProviderUrl(args);
+    const turnkeyPublicKey = await this.initIframeStamper();
+    const oauthParams = args;
+    const providerUrl = await this.getOauthProviderUrl({
+      oauthParams,
+      turnkeyPublicKey,
+      oauthCallbackUrl: this.oauthCallbackUrl,
+    });
     const popup = window.open(
       providerUrl,
       "_blank",
@@ -554,85 +556,6 @@ export class AlchemySignerWebClient extends BaseSignerClient<ExportWalletParams>
         clearInterval(checkCloseIntervalId);
       };
     });
-  };
-
-  private getOauthProviderUrl = async (args: OauthParams): Promise<string> => {
-    const {
-      authProviderId,
-      isCustomProvider,
-      auth0Connection,
-      scope: providedScope,
-      claims: providedClaims,
-      mode,
-      redirectUrl,
-      expirationSeconds,
-    } = args;
-    const { codeChallenge, requestKey, authProviders } =
-      await this.getOauthConfigForMode(mode);
-    if (!authProviders) {
-      throw new OAuthProvidersError();
-    }
-    const authProvider = authProviders.find(
-      (provider) =>
-        provider.id === authProviderId &&
-        !!provider.isCustomProvider === !!isCustomProvider
-    );
-    if (!authProvider) {
-      throw new Error(`No auth provider found with id ${authProviderId}`);
-    }
-    let scope: string;
-    let claims: string | undefined;
-    if (providedScope) {
-      scope = addOpenIdIfAbsent(providedScope);
-      claims = providedClaims;
-    } else {
-      if (isCustomProvider) {
-        throw new Error("scope must be provided for a custom provider");
-      }
-      const scopeAndClaims = getDefaultScopeAndClaims(authProviderId);
-      if (!scopeAndClaims) {
-        throw new Error(
-          `Default scope not known for provider ${authProviderId}`
-        );
-      }
-      ({ scope, claims } = scopeAndClaims);
-    }
-    const { authEndpoint, clientId } = authProvider;
-    const turnkeyPublicKey = await this.initIframeStamper();
-    const nonce = getOauthNonce(turnkeyPublicKey);
-    const stateObject: OauthState = {
-      authProviderId,
-      isCustomProvider,
-      requestKey,
-      turnkeyPublicKey,
-      expirationSeconds,
-      redirectUrl:
-        mode === "redirect" ? resolveRelativeUrl(redirectUrl) : undefined,
-      openerOrigin: mode === "popup" ? window.location.origin : undefined,
-    };
-    const state = base64UrlEncode(
-      new TextEncoder().encode(JSON.stringify(stateObject))
-    );
-    const authUrl = new URL(authEndpoint);
-    const params: Record<string, string> = {
-      redirect_uri: this.oauthCallbackUrl,
-      response_type: "code",
-      scope,
-      state,
-      code_challenge: codeChallenge,
-      code_challenge_method: "S256",
-      prompt: "select_account",
-      client_id: clientId,
-      nonce,
-    };
-    if (claims) {
-      params.claims = claims;
-    }
-    if (auth0Connection) {
-      params.connection = auth0Connection;
-    }
-    authUrl.search = new URLSearchParams(params).toString();
-    return authUrl.toString();
   };
 
   private initIframeStamper = async () => {
@@ -723,38 +646,6 @@ export class AlchemySignerWebClient extends BaseSignerClient<ExportWalletParams>
     const nonce = getOauthNonce(publicKey);
     return this.request("/v1/prepare-oauth", { nonce });
   };
-
-  private getOauthConfigForMode = async (
-    mode: OauthMode
-  ): Promise<OauthConfig> => {
-    if (this.oauthConfig) {
-      return this.oauthConfig;
-    } else if (mode === "redirect") {
-      return this.initOauth();
-    } else {
-      throw new Error(
-        "enablePopupOauth must be set in configuration or signer.preparePopupOauth must be called before using popup-based OAuth login"
-      );
-    }
-  };
-}
-
-function resolveRelativeUrl(url: string): string {
-  // Funny trick.
-  const a = document.createElement("a");
-  a.href = url;
-  return a.href;
-}
-
-/**
- * "openid" is a required scope in the OIDC protocol. Insert it if the user
- * forgot.
- *
- * @param {string} scope scope param which may be missing "openid"
- * @returns {string} scope which most definitely contains "openid"
- */
-function addOpenIdIfAbsent(scope: string): string {
-  return scope.match(/\bopenid\b/) ? scope : `openid ${scope}`;
 }
 
 /**
