+++ account-kit/signer/src/utils/resolveRelativeUrl.ts
@@ -1,0 +1,6 @@
+export function resolveRelativeUrl(url: string): string {
+  // Funny trick.
+  const a = document.createElement("a");
+  a.href = url;
+  return a.href;
+}
