+++ site/pages/reference/account-kit/signer/classes/BaseSignerClient/getOauthProviderUrl.mdx
@@ -1,0 +1,56 @@
+---
+# This file is autogenerated
+
+title: getOauthProviderUrl
+description: Overview of the getOauthProviderUrl method
+---
+
+# getOauthProviderUrl
+
+Returns the authentication url for the selected OAuth Proivder
+
+## Import
+
+```ts
+import { BaseSignerClient } from "@account-kit/signer";
+```
+
+## Usage
+
+```ts
+
+cosnt oauthParams = {
+authProviderId: "google",
+isCustomProvider: false,
+auth0Connection: undefined,
+scope: undefined,
+claims: undefined,
+mode: "redirect",
+redirectUrl: "https://your-url-path/oauth-return",
+expirationSeconds: 3000
+};
+
+const turnkeyPublicKey = await this.initIframeStamper();
+const oauthCallbackUrl = this.oauthCallbackUrl;
+const oauthConfig = this.getOauthConfig() // Optional value for OauthConfig()
+const usesRelativeUrl = true // Optional value to determine if we use a relative (or absolute) url for the `redirect_url`
+
+const oauthProviderUrl = getOauthProviderUrl({
+oauthParams,
+turnkeyPublicKey,
+oauthCallbackUrl
+})
+
+```
+
+## Parameters
+
+### args
+
+`GetOauthProviderUrlArgs`
+Required. The Oauth provider's auth parameters
+
+## Returns
+
+`Promise<string>`
+returns the Oauth provider's url
